mon oconf push: Use mon oconf reload when restarting peers and pollers
This largely avoids the issues with config changes triggering mass
spamming of the report_data table when the merlin daemon loses track
of which nodes it's already inserted, and limits stable hosts and
services to a single insert per week, which is actually pretty good,
since it means we never have to drop extremely far back when finding
initial states for objects we want to create reports for.

This fixes issue #5056 for the "config reload" case. During normal
operations, that's the really interesting part, so we'll mark that
issues as fixed with this patch.

Signed-off-by: Andreas Ericsson <ae@op5.se>