apps/test: Add test that currently crashes nagios
Because we replaced our own downtime deletion code with the nagios
built-in code, which was racy and didn't work, we started crashing
whenever propagating triggered downtimes were deleted. This change adds
a test that shows the problem to mon test dist, so we see when it is
fixed.

The error is displayed as a connection error to livestatus, which looks
confusing at first, but once you git blame the code that's running,
you'll realize that mon test couldn't connect to livestatus because the
nagios you are trying to reach has expired and gone to meet it's maker.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>