install-merlin: Stop chmod'ing our execs
If gcc doesn't produce executable files we're in serious trouble
anyway, and the script files are all chmod'ed +x when they're put
in their proper place, so the only purpose these three lines served
was to aggravate testers running this from a git checkout and getting
init.sh chmod'ed to +x in the git checkout. Not cool.

Signed-off-by: Andreas Ericsson <ae@op5.se>