module: Send CTRL_ACTIVE volley when we get updates from neighbours
This means we'll be blasting the net pretty hard the first few
seconds of operations, but during normal operations this doesn't
appear to have any detrimental effect what so ever. It does however
make node discovery and negotiation go a lot faster, so tests run
all that much faster too, and checks are less likely to bounce
back and forth between systems, which is nice.

Signed-off-by: Andreas Ericsson <ae@op5.se>