module: Byebye control.c. Hello pgroup.c
The pgroup "api" implements peer/poller groups, which are highly
useful when tracking a collection of nodes that in turn keep
track of a specific set of hosts or services. We primarily use
them to, with great accurracy, track how many checks each node
should run and cache that information for speedy access in case
nodes come and go.

This means a fairly major architectural overhaul, but since the
control.c thing was a bit of a hack from the start that dwindled
(as hacks do) over time and this provides us with very detailed
stats and tracking of checks and configuration errors, this is
definitely a good move.

As a neat bonus, peer id negotiation is now cut down to about two
seconds for the most part, which is a significant improvement over
the earlier "10 seconds or more". That's a side-effect of using a
consolidated api that tracks and triggers on changes in all peer
groups we know of.

Signed-off-by: Andreas Ericsson <ae@op5.se>