Absolutize and retain the path to our configuration file
For the module, we do this by taking the path relative to the
nagios.cfg file, which is how Nagios does things nowadays, so
it's neat to maintain conformity.

For the daemon, we take the path relative to our current directory.

This makes it possible to print the config file name we're using in
case we run into odd problems, which comes in handy when running
multiple instances on the same system (as is the case when we run
tests).

Signed-off-by: Andreas Ericsson <ae@op5.se>