module: Wait for peers and pollers to connect before we get busy
When we first launch and get our ipc connection going, we used to
start monitoring things immediately, potentially launching a bunch
of notifications that our fellow peers and pollers were supposed to
take care of.

With this patch, we wait a brief moment (up to 10 seconds or until
all our peers and pollers are connected, whichever comes first), and
only poll for input through the iobroker while we're waiting.

Using the iobroker is actually quite neat here. It means we don't
have to set up our own polling infrastructure, so the code is kept
at a bare minimum, and it also means the query handler works
normally during the brief pause while we await the friendly greetings
from our neighbour nodes.

Some preliminary testing shows that it now takes, on average, 0.9
seconds for the stagger-started systems of 'mon test dist' to find
all their neighbours and kickstart monitoring. We've put the upper
limit at 10 seconds, which should be fine given that all other
infrastructure will work normally during the break.

While at it, we lower the connect time for mon test dist from 10 to
3 seconds, and tests now pass "better" than they used to. Quite nice,
imnsho.

This will most likely fix issue #6743.

Signed-off-by: Andreas Ericsson <ae@op5.se>