mon test dist: Make tests selectable
It's really, really annoying to try to run tests for a specific
thing when it takes a good minute or two to get to the tests one's
interested in.

With this patch, we recognize all functions named 'test_' in the
fake_mesh class and make them available as selectable tests. By
default, all tests are selected, which matches the earlier
behaviour.

--list (or --test=list) can be used to list all available tests. It's
up to the caller to figure out that 'acks' depend on 'passive_checks'
though.

This means a few functions had to change names, so the patch looks
bigger than it is. It seems to work ok though, so let's go with
this.

You're welcome, Robin.

Signed-off-by: Andreas Ericsson <ae@op5.se>