module: Revamp node check detection somewhat
First off, we utilize a macro in module.h to do the actual math,
so we can run tests til our heart's content. We also make the
should_run_check() macro a function and make it take an unsigned
int (for object id) rather than taking the reference from the
object pointer in the macro. This lets us add debugging stuff
if we're so inclined.

Lastly, we change all callsites so the code actually compiles,
because I've been told that's Sound And Sensible.

Signed-off-by: Andreas Ericsson <ae@op5.se>