module: Don't enforce coredumps and the chdir() home thing
Since quite a while back, Nagios already does the chdir() home
stuff, but with a much cleverer scheme than we do (such as only
changing directories if it's running as a daemon), and since a
slightly shorter but still long enough time Nagios dumps core by
default. If users configure Nagios not to, we shouldn't change
their minds for them.

Signed-off-by: Andreas Ericsson <ae@op5.se>