spec: Make suse compatible, move mon include dir (again)
This should thus resolve bug #6505.

This also moves the mon command and other misc merlin lib files from
libexec to the systems libdir.

FHS says the lib directory is just as correct as libexec. Suse and, if
my memory serves correctly, debian says "screw libexec, we're goin' with
lib all the way!". Using /usr/lib{,64}/merlin instead of
/usr/libexec/merlin thus reduces the differences between rhel and suse
for us, which should be a good thing.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>