Revamp how we consider 'self'
The module in particular is riddled with places where 'self' and
ipc.info is almost, but not quite, interchangeable, and we try
pretty damn hard to keep our head straigt with regards to which
of the two separate structs we update.

Since the introduction of the pgroup concept, this has gotten
worse, as pgroups want to treat all nodes the same (with the
exception of connectedness, since "ipc" must always be considered
alive for the world to make any sense at all).

We can't really remove the self struct, since the daemon has to
stash the module's info somewhere. ipc.info is currently used for
this. However, the issue is microscopic for the daemon as it doesn't
really care what info other nodes send. We take care of that by the
rather simple expedient of changing 'self' to a pointer, which we
point to ipc.info in the module and to a newly introduced variable
'merlind' in the daemon.

A neat effect of this simplification is that we now report correct
numbers for how many checks each node has executed, which provides
the base for fixing issues #6446, #6745, #6712 and #5818.

Signed-off-by: Andreas Ericsson <ae@op5.se>