module: Stop bouncing passive checkresults back to the network
Passive checks are always sprayed from the originating node and
should be scheduled and handled identically on all nodes that
receives the result, so there's no need to volley it back to the
network.

This makes passive checks appear to be executed on each node (or
rather, the node that gets the "nodeinfo" query), which solves
at least a part of issue #5818.

Signed-off-by: Andreas Ericsson <ae@op5.se>