module: Fix sigsegv for clusters with master-peers > pgroup-nodes
A typo from the initial implementation of pgroups caused us to
dump core on startup when we had a poller-group with fewer nodes
than there were top-level peers. The problem was two-fold. First
off, we over-allocated the space used for the nodelist. Secondly,
we used the alloc-counter to iterate over the nodes, causing an
illegal access on the pgroup->total_nodes + 1 node encountered.

This patch fixes both issues, providing belts and suspenders for
the problem.

Signed-off-by: Andreas Ericsson <ae@op5.se>