Remove nebtest
It's pretty much obsolete since we now have 'mon test dist' and
doesn't provide the value it should for the maintenance cost it
incurs. The functionality would be much better off if it was
built into Nagios anyway.

This also means we can remove nagios-stubs.h, which is a serious
pain to maintain (which is probably the reason why it hasn't
been).

Signed-off-by: Andreas Ericsson <ae@op5.se>