mon test: Revamp mesh destruction
Using a non-class function to shut down the mesh sort of made
sense about 3000 revisions ago when it was the primary signal
handler, but nowadays it just leads to more typing and variables
one has to remember to pass for the function to work properly
without throwing exceptions.

This way is far cleaner, and allows for critical tests to call
the shutdown functions themselves without having to go the
completely ass-backwards way of an external function that uses
a global parameter.

Signed-off-by: Andreas Ericsson <ae@op5.se>