mon test dist: Add daemon status tests
This test checks that the daemons each instance has spawned
are still reachable via kill(pid, 0). Failure to signal the child
means it has died.

It's worth noting that it's possible (but unlikely) that a
different process that we can signal has taken over the pid,
which will lead to false positive result of this particular
test.

A better test would be to catch SIGCHLD and set the status of
each process correctly, but that makes restart tests and process
tracing exceptionally cumbersome to handle, so I didn't really
bother.

Signed-off-by: Andreas Ericsson <ae@op5.se>