module: Fix pgroup allocation
This patch half-reverts af1d80576477fd50a30d9f69392be97ca6501fca.

It appears that the suspenders from that patch was mainly used to
hang ourselves with in case all our pollergroups didn't have the
same amount of peers as ipc.pgroup has, which is pretty unfortunate
since that was exactly the bug the original patch was intended to
fix in the first place. So it's official; I suck at testing.

Signed-off-by: Andreas Ericsson <ae@op5.se>