mon check distribution: Fix the horrible brokenness we had recently
Just that. It was horribly broken earlier. With this patch it works
implicitly by reporting the numbers we get from the merlin module,
although the window we allow for checks is now ridiculously high,
it should cover pretty much all cases, with the possible exception
of a multitude of passive checks.

However, if we're going down that route, we shouldn't blame the
mon check thing, but rather the merlin module (since that's the
one who's now reporting the numbers we use).

This fixes issue #6712.

Signed-off-by: Andreas Ericsson <ae@op5.se>