daemon: Do sighandling properly
By "properly" I mean "catch the signal, mark it in a variable and
exit the hell out of the handler asap".

This closes the microscopic window where two real-time signals are
queued almost simultaneously, there is no active database connection,
use_database = 1 and db_track_current = 1 and we get the second
signal while attempting to connect to the database in order to send
our query.

It's a damn narrow window, but it's still a good idea to get this
stuff sorted out so we never leave coredumps behind.

While at it, make sure we exit with return code 0 when we catch
sigterm or sigint.

Signed-off-by: Andreas Ericsson <ae@op5.se>