mon test dist: Revamp aliveness tests for better granularity
With this patch, we implement the following rules:
* "exit nicely" means "with return code 0" and not due to a signal.
* A dead deamon causes a failure no matter what.
* Coredumps are always fails (and are now printed in red)
* Signals can be ok, if we've told the test_alive() function that
  some are.
* Any kind of fallthrough causes a failure

This lets us make sure daemons shut down properly without dumping
core or misbehaving in other similar manners.

While at it, we fix a syntax error in the test_alive() function so
we can actually use the kwargs we earlier only pretended to handle.

To keep tests running nicely, we have to increase nagios' deathtime
somewhat. With the old deathtimer, we would sometimes get test
failures due to too quick measuring.

Signed-off-by: Andreas Ericsson <ae@op5.se>