mon test dist: Fix connection tests with more than one poller-groups
Earlier we compared the counted connections with the total number
of nodes in the entire mesh, which was a bit off, as separate poller
groups don't know about each other. Fortunately, each node already
trackes (and counts) the number of neighbours it has, so the fix is
an easy one.

Signed-off-by: Andreas Ericsson <ae@op5.se>