module: Get rid of the node action handler
We do treat the ipc node specially, so that one remains, but as
we have to force the peer id assignment on every new CTRL_ACTIVE
packet anyway, the node-action thing was reduced to handling
disconnects, which it didn't really do that great a job of for
anything but the ipc node itself.

This patch fixes the peer-id assignment discrepancies once could
sometimes see for poller groups, while removing quite a lot of
code in the process. Neato.

Signed-off-by: Andreas Ericsson <ae@op5.se>