mon check distribution: Stop throwing exceptions for critical states
Previously we used to try to access the 'should' variable to see
how many checks we really should check, but with a recent patch
we stopped populating it.

Since that patch also modified the logic so that the state should
only return non-OK when something's really off, the bug went
unnoticed.

This patch appears to fix the problem though.

Signed-off-by: Andreas Ericsson <ae@op5.se>