mon node tree: Print ascii tree (map) of merlin nodes
A script that shows the current setup of nodes by generating a map
made using ascii characters.

The map represents the current view of the node network in the
eyes of the node that the script is executed on.

For instance, this means that the network might look a whole lot
bigger from a master node's point of view than what it does for
one of its pollers.

Illustrating the nodes like this helps anyone who is administrating
the system getting a quick understanding of what sort of setup that
they are dealing with.

Additionally, this makes it easy to spot simple, but fatal, node
misconfigurations such as a node being configured as e.g. a peer
instead of a master.

Signed-off-by: Philip Eklöf <peklof@op5.com>