mon oconf nodesplit: Handle absolute cfg_dirs
When exposed to an absolute cfg_dir, we would try to access something
like /opt/monitor/etc//opt/monitor/etc/configfile.cfg - as that file
didn't exist, we would ignore it. This meant our timestamps was always
for the latest file explicitly named in the nagios conf, which caused
infinite restart loops.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>