test: Skip validation when using "mon test dist" to generate config
"mon test dist", which is a tool for generating clusters, can also be
used as a tool for generating configs. Weird as that may be, what's even
weirder is that it will loudly refuse to generate configs without poller
groups and too few masters, even when I really want a nice, neat,
large-ish config without too much typing.

With this patch in place, I will be allowed to generate master-only
configs.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>