daemon: Be consistent with address resolution errors
Previously we used to bail out on bogus addresses inside node
compounds, like "address = foobarlollipop", but not for implicit
addresses where we used the node name as the address. This patch
corrects that, so we bail out early on resolve errors no matter
how they are configured.

This fixes issue #6883.

Signed-off-by: Andreas Ericsson <ae@op5.se>