mon test dist: Make downtime tests a bit more thorough
Downtime really consists of three parts; The downtime depth marker
in the object, the downtime object itself, and a downtime comment
for the downtime entry itself. In order to fully verify that
downtime handling works properly, we need to make sure all three
parts are handled correctly and are kept in sync.

To facilitate this type of testing, which has to be added to a lot
of places, we introuce a simple helper to the 'mesh' class which
tests all elements of downtime. Since we previously had a lot of
duplicated code to test the same thing, the net result is actually
a small reduction in codesize.

Signed-off-by: Andreas Ericsson <ae@op5.se>