mon test dist: Make valgrind logs unique
Previously Nagios and Merlin would mash their valgrind logs into
a single file. That was stupid, and is corrected with this patch.
We also take care of restart tests, which should preserve the old
valgrind logs and open a new one instead. The logs are now numbered
sequentially per daemon so we can track sort of easily which
invocation of the program a particular log pertains to.

Signed-off-by: Andreas Ericsson <ae@op5.se>