module: Set inactive node's assigned checks to 0
This actually fixes an illegal memory access, since we earlier
would attempt to read beyond our allocated storage when accessing
the node assignment table (which is as small as it can be).

The bug was hidden by the fact that the data read was still on the
heap and implicitly initialized to 0 by the kernel, so the code
actually did the right thing when read-smashing its own heap. Neat,
and actually quite cute, but bad for valgrind errors and it will
fail when using memory markers to detect bad memory accesses, so
let's fix it.

Signed-off-by: Andreas Ericsson <ae@op5.se>