mon test dist: Make this a bit more valgrind friendly
Earlier we used to send a SIGKILL about half a nanosecond after we
sent SIGTERM to our daemons, which means the interesting part of
any valgrind info was left out from the logs.

To avoid that, we wait up to 5 seconds, but since I'm impatient as
hell, we try to wait() for the children who were just recently sent
a SIGTERM. In practice, this means the exit is delayed by about 50ms,
which is tolerable.

Signed-off-by: Andreas Ericsson <ae@op5.se>