Janitorial: Stop shadowing variables and parameters
Detected by using more stringent warning options to gcc, this
actually fixes a real bug, in which PROGRAM_STATUS_DATA and
CONTACT_NOTIFICATION_METHOD_DATA events were marked as coming from
the wrong node when entered into the database.

Even though the database is destined for obsoletion and removal, this
is a class of bugs that's relatively easy to avoid, so let's fix
these instances now before we enable a bunch of additional warnings.

Signed-off-by: Andreas Ericsson <ae@op5.se>