Janitorial: Fix prototypes
-Wstrict-prototypes complains when functions not accepting any
arguments lack the (void) marker. When present, it allows the
compiler be certain that the function really doesn't take any
arguments, rather than having to assume that it's an ansi c style
forward declaration and build the code accordingly. For this
particular case, it doesn't matter all that much, but since we
want to enable those warnings, we need to fix this particular
warning first.

Signed-off-by: Andreas Ericsson <ae@op5.se>