Janitorial: Fix a missing initializer warning in db_wrap
When we added the conn_str option to the db_conn_parameter struct,
we forgot to add an initializer for it in the 'empty' struct. This
isn't a real bug, since structs with partial initializers have to
get all their remaining variables explicitly initialized to zero (or
NULL, as the case may be), but since we use a lot of structs as lists
with constant initializers, we really should get this warning fixed
so it can detect real bugs in the future.

Signed-off-by: Andreas Ericsson <ae@op5.se>