Janitorial: Remove the 'const' qualifier for the sql_port function
Having it there makes no sense, as it's not used in loops or anywhere
else where the call can be omitted because the compiler already
knows the answer, and it returns by value rather than by reference,
so the caller really can modify it to its hearts content.

It also silences a warning about ignored const qualifiers, which is
a Good Thing(tm).

Signed-off-by: Andreas Ericsson <ae@op5.se>