Add experimental waf support for Merlin
Still incomplete, as we don't do variable substitution and things
like that, but at least it detects if libnagios is available and
which database libraries we can build against.

Note that this isn't used by default, and isn't intended to be
used by default until it works better than the current Makefile
+ install-merlin.sh combo, but at least it gives people a way
to play with it that's sort of half-official.

Signed-off-by: Andreas Ericsson <ae@op5.se>