mon oconf: Try a bit harder to avoid doing work
Because laziness is a virtue. Or something.

On a more serious note, this fixes a real (although impotent) bug
which earlier caused the run_param() function to be called even
when we had no work for it to do. There's an explicit check to
see if all files are cached, but it checked the length of the wrong
list, so it was broken. This fixes it.

Signed-off-by: Andreas Ericsson <ae@op5.se>