mon oconf: Add support for using nagiosplit to separate configs
A "quick" test shows it to be a bazillion times faster than using
the home-hacked python code, and since we can run several at a time,
we're able to fully utilize SMP capabilities of our target systems.

It's possible that some very odd configurations are actually faster
to split with the old code (if the pollers are extremely small and
the configuration is medium-sized), but I doubt it.

One nice effect of this is that we now include all hosts in the
hostgroups listed on the 'hostgroups' statement in merlin.conf,
which the old code didn't do. One bad effect of this is that we
no longer include the perfdata commands, which the old code did.
Fixing the helper to handle that properly shouldn't be hard though,
but it's pending a slight rewrite of the perfdata handling routines
in Nagios, which will let the commands be parsed by the main config
parsing routine instead of in a separate place.

The timings below are taken with a config containing 3000-ish hosts
and roughly 25000 services. It's big, but not extraordinarily so.
There are three groups of pollers (peered pollers still get their
config from hardlinks, which are near-instantaneous).

Without helper
$ /usr/bin/time mon oconf nodesplit --force
/usr/bin/time mon oconf nodesplit --force
/var/cache/merlin/config/poller1.cfg created with 3219 objects for hostgroup list 'NET,DAI,ADC'
/var/cache/merlin/config/poller2.cfg created with 39851 objects for hostgroup list 'WIN,UNIX'
/var/cache/merlin/config/poller3.cfg created with 16444 objects for hostgroup list 'FLS,ESS'
270.00user 0.67system 4:31.81elapsed 99%CPU (0avgtext+0avgdata 536960maxresident)k
0inputs+173912outputs (0major+139918minor)pagefaults 0swaps

With helper
$ /usr/bin/time mon oconf nodesplit --force --use-helper
3150 objects cached to /var/cache/merlin/config/poller1.cfg
16395 objects cached to /var/cache/merlin/config/poller3.cfg
39704 objects cached to /var/cache/merlin/config/poller2.cfg
9.40user 0.31system 0:03.80elapsed 255%CPU (0avgtext+0avgdata 148488maxresident)k
0inputs+124336outputs (0major+153999minor)pagefaults 0swaps

From 270 seconds to 9.4. Not too shabby.

This probably fixes issue #6970, although more thorough correctness
checks and testing would be good.

Signed-off-by: Andreas Ericsson <ae@op5.se>