module: Send has_been_checked with all updates
Commit 3990f01f915cea0520ce0fd45f325c5d8002bfbe tried to achieve
this properly (and made it at least auto-healing), but the patch
was backwards and caused us to read the has_been_checked value
but never set it. This patch amends it so we set the value
properly but never ship it. Instead, the receiving end has to
rely on getting a checkresult for the object in question for
the flag to get set.

This should fix the very-hard-to-test issue #7020.

Signed-off-by: Andreas Ericsson <ae@op5.se>