mon test dist: Allow a more flexible specification of poller-groups
Using poller-groups where all groups are the same size makes little
sense, since some bugs only show up when we have poller-groups that
are dissimilar size or aren't the same size as there are masters.

This patch removes the --pollers-per-group and --poller-groups args
and replaces it with a --pgroups argument which takes a comma-separated
list of integers. The length of the list represents the number of
pollers, and the numbers represent the number of pollers per group,
so the new default [5,3,1] means "add 3 poller-groups. The first
should have 5 pollers, the second 3 and the last one just a single
one".

This gives a lot better coverage of actual scenarios, which is nice.

Signed-off-by: Andreas Ericsson <ae@op5.se>