mon test dist: Autogenerate hosts and services cleverly
It's nifty to have a number of services that divides evenly with
the number of hosts in the group, since that makes it trivial to
visually ensure that all nodes run their assigned checks. It's
also nifty to have an uneven number of hosts, since that means we
don't make it too easy for the config splitting algorithm, as one
node will have to do more work than the others, even if it's just
one extra hostcheck.

This patch ensures that that happens by setting the default value
for num_hosts and num_services to False and letting that trickle
through to the fake_peer_group.create_object_config() method,
which handles all the logic involved.

Signed-off-by: Andreas Ericsson <ae@op5.se>