pgroup: Assign poller-group checks based on poller's host id's
Since pollers' objects don't get the same object id's as the master,
we must use the counter for the number of hosts the poller have
assigned to it to figure out which of the pollers in the group
will run which number of hosts and services.

Assuming the following conditions, we'd otherwise report the wrong
number for assigned checks for pollers:
* The master group has X objects not distributed to pollers
* Out of X, Y objects are lexicographically sorted before all objects
assigned to the affected poller-group.
* Y is a number not evenly divisable with the number of masters
peers in the master-group.
* The poller group contains more than one poller.

This patch fixes it neatly, however, so tests pass once again.

Signed-off-by: Andreas Ericsson <ae@op5.se>