mon test dist: Avoid os.WNOHANG when stopping daemons during shutdown
Doing so means we can get test-failures during shutdown, which aren't
very interesting. We still want to see all exceptions (like sigsegv
and other annoying things), but we really don't care if a particular
process has managed to exit already or not.

While at it, fix the exception case where we have no children to mean
success if we expect daemons to be dead, and handle it gracefully (ie,
without a stacktrace) if we expect them to be alive.

Signed-off-by: Andreas Ericsson <ae@op5.se>