mon node status: Detect and print obvious config errors
It happens that users get their configuration wrong, or fail to
set up configuration sync the way it's supposed to be done. Since
the query-handler support was introduced in Merlin, it's really
easy to see when such things are wrong, so let's do just that and
alert the user to it, in loud, bright colors.

Signed-off-by: Andreas Ericsson <ae@op5.se>