test: Add more wireframe and dummy data
We should now be able to start writing some tests by decoding the
packages that arrive at the ipc_send_event function. This allows us to
verify that any transformations/mutations done to objects are correct.
Of course, it also enables checking that mutations are NOT done we don't
want them to be.

Signed-off-by: Anton Lofgren <alofgren@op5.com>