mon test dist: Use progressive tests for active check tests
Running for two minutes and then checking to see if the checks have
bounced to their right node is sort of useless, since things are
completed way sooner than that. This patch let's the suite move on
to the next subsuite as soon as all checks have been reassigned to
their proper node.

Signed-off-by: Andreas Ericsson <ae@op5.se>