mon test dist: Add test_parents()
The tests aren't actually fully automated yet, but with this test we
run checks in a predetermined order and the results are verifyable
manually. Adding automagic verification shouldn't be all that hard,
but that can be done later (since not all tests would pass anyway).

This provides the basis for testing issue #6552.

Signed-off-by: Andreas Ericsson <ae@op5.se>