mon node status: Fix inconsistencies in how we print durations
Since the code is a legacy one, some of the strings we printed
in case a node was down or unresponsive were pretty weird. This
patch makes us print "unknown" for all of them when it's actually
unknown. That looks sane, since we print the durations/timings
after a "label:" marker.

This also makes it easier to script against the mon node status
output, even though that's sort of frowned upon, as it's intended
for human consumption only.

Signed-off-by: Andreas Ericsson <ae@op5.se>