mon test dist: Increase timings somewhat
It appears that our poor CI servers are completely useless when it
comes to performing properly, so they're slow in forwarding events
and everything related to it. With this patch we increase the wait
time somewhat for most tests, so our underdimensioned test servers
actually have time to complete the tests.

Signed-off-by: Andreas Ericsson <ae@op5.se>