mon log: Make people aware that "--help" is sometimes nifty to use
This is true in particular for "mon log import" and "mon log show",
which both use sophisticated backend programs with a multitude of
options. We really don't want to duplicate those options here, or
invent a new way to pre-load particular commands of python-modules
with some extra helptext, but this patch is cheap, works and will
point users in the right direction.

This fixes issue #6942.

Signed-off-by: Andreas Ericsson <ae@op5.se>