mon node remove: Fix removing multiple nodes in one command
Since we use a sed-line to remove the offending compound from the
configuration file, we must re-parse the entire file again for each
node we're supposed to remove so we get the correct lines for the
compounds even after we've removed an earlier node.

A neater solution would've been to sort them in order of "last in file"
and then just parse it once, but that's more trouble than this simple
patch, so I'm not gonna do that.

Signed-off-by: Andreas Ericsson <ae@op5.se>