ocimp: Fix downtime junk with new nagios'es
The previous patch was a tad naive in that it didn't change the
offsets used to address author and comment fields, so not everything
was parsed (and inserted) properly with just the earlier patch. This
does things properly, albeit with a microscopic slowdown as we now
do strcmp() for every variable (beyond the first two) that we parse.

Signed-off-by: Andreas Ericsson <ae@op5.se>