test: Wrap notification log discovery in ptest
Our sles build server cached the writes for long enough for the test
runner to decide that the file didn't exist and move on.

In fact, sles having some kind of disk buffer tuned to throughput and
centos having that same buffer tuned towards interactivity could explain
all sorts of weird timing issues that's shown up on sles but not centos.
Interesting.

The obvious patch is to just flush&fsync in the mark command.
Surprisingly, that doesn't work, so do something else.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>