module: Register check results as normal check results
Previously we used to snapshot the state of an object and, having
received that object, we would copy it to our version of it, in-place.
That meant we always had the exact same data, but got none of the same
triggers on multiple systems. That, in turn, lead us to sometimes
missing out on parent/child relations, or host/service relations and
thus sending out the wrong kind of notification.

With this patch, each checkresult that comes in from any merlin node
is considered equal to all checkresults that arrive from any of our
workers. Nifty, and very practical, as we now get our dependency
checks done properly.

The downside is that we'll sometimes block or mistreat hosts and
services that currently get run as part of the dependency checks,
so while we don't exactly *need* the very latest Nagios version we
can lay our hands on, it sure would be nice in order to avoid a
certain degree of weirdness with check scheduling when we hit a
snag in the cluster.

Signed-off-by: Andreas Ericsson <ae@op5.se>