module: Support configuring which node notifies and which doesn't
This is the first patch towards fixing issue #5785, and it needs
to be documented. When "notifies = no" is set inside a poller's
compound in that poller's master's merlin.conf, the idea is that
the master will send notifications that the poller would otherwise
have sent. The poller must also have "notifies = no" inside its
"module" compound in its merlin.conf.

To recap,
On the master node:
poller poller-01 {
	address = 10.11.12.13
	hostgroups = poller-01-hosts
	notifies = no
	# other vars...
}

On the poller "poller-01":
module {
	notifies = no
	# other vars...
}

If this is misconfigured so that only the poller is blocked from
sending notifications, but the master doesn't know about it, no
notifications will be sent for checks executed on or received by
the poller.

If this is misconfigured so that the master thinks notifications
are blocked on the poller when they really aren't, one will receive
duplicate notifications from checks executed on or received by the
poller node.

Signed-off-by: Andreas Ericsson <ae@op5.se>