module: Block notifications we're not supposed to send
This quite small patch makes it possible for us to adhere to the
new config options we've added and block notifications on pollers
that still have masters connected if the pollers are configured that
way. If the masters are configure properly as well, they will instead
send the notifications themselves.

This fixes issue #5785.

Signed-off-by: Andreas Ericsson <ae@op5.se>