module: Fix transferrals of host and service checks
Two separate bugs had snuck in there, and this fixes both of them by
making sure we handle host and service checks identically.

First off, we never transfer passive checkresults as checkresult events
across the network, but instead mark them as MAGIC_NONET. That fixes
the tests for service checks, which got confused regarding assignment
when they were received as passive checks followed by active checks
which should be rescheduled, but with the exact same results.

The second bug we solve is that passive host checks now once again
register in the report database, since we only mark them with the
MAGIC_NONET flag instead of returning 0 whenever we encounter such
a checkresult.

Signed-off-by: Andreas Ericsson <ae@op5.se>