mon: Fix return codes of main syntax help text
If mon is run with a bad set of arguments or anything similar, it
should definitely exit with a return code above 0 and show its
syntax help text. But if a user specifically asks for the help text
by running e.g. "mon --help", then no error has occurred and the
exit code should be set to 0.

Signed-off-by: Philip Eklöf <peklof@op5.com>