module: Add code to track which node is responsible for a check
We've always had this, but only for peers, so things got tricky for
the pollers. In order to do this properly, we need to implement
object id conversion tables, so our object id's map to the ones used
by the poller(s) in question. Without it, there's no way we can say
which node should run which check, and we need to know that in order
to report on errors when checks get run by the wrong node.

Signed-off-by: Andreas Ericsson <ae@op5.se>