module: Stop sending network received events back to the network
It causes unnecessary traffic and fouls up the check distribution
tracking.

This also fixes the weird and faulty checks for host and service
selection process which was in place earlier, and makes sure we
only ever do the selection lookup process if we're really not
going to send the packet over the wire.

Signed-off-by: Andreas Ericsson <ae@op5.se>