module: Augment check tracking so we can warn about bad assignments
This builds upon the recent changes to pgroup.c, which allowed us to
determine which poller in a group of pollers was responsible for a
check by adding an error messages to the logs in case we get a check
result from the wrong node.

We only warn about active checks though, since passive checks always
get sent as checkcommands rather than as checkresult events.

Signed-off-by: Andreas Ericsson <ae@op5.se>