module: Remove some excessive logging
Logging every host we look at when doing pgroup mapping is just stupid
and detracts from the valuable parts, so let's stop doing that.

While at it, move up the bitmap_set() operation to tag which hosts
are handled by the pgroup to where we check if it's already set, so
we don't have to go looking for it.

Signed-off-by: Andreas Ericsson <ae@op5.se>