module: Count inheritance objects separately from assigned objects
When an entire poller group goes offline, the masters don't
necessarily inherit items in the same peer-order that the pollers
used to distribute them (so, for instance, the poller's host #13
may well be our host #244, and should thus be distributed to a
different peer for the masters than it should for the pollers).

This should make us correctly calculate assigned objects when one
or more groups of peered pollers are offline.

Signed-off-by: Andreas Ericsson <ae@op5.se>