module: Let ipc.pgroup inherit assigned checks from offline pgroups
The totals only, but this should provide more accurate and less
confusing numbers when we're looking at a master node with one or
more poller-groups attached where those pollers are offline.

Earlier, the number of executed host or service checks could in
some cases be greater than the number of hosts assigned to the
(master) peer group, which was wildly annoying and looked just
plain wrong. This patch should correct that, allowing for pollers
to come and go as they see fit while we maintain correct check
accounting.

Signed-off-by: Andreas Ericsson <ae@op5.se>