module: Fix assignment calculation for ipc.pgroup
The ipc peer-group is special, since we by default handle exactly
everything when we have no pollers online, so we must *start* by
setting our "assigned" to everything, and then deduct the number
of poller-assigned hosts and services for poller-groups that have
at least one poller online.

This doesn't handle the case where "takeover = no" is set, in which
case the pollers are still responsible for the objects they watch
even if no pollers are online and active, but that should be a rare
case and will anyways result in an error from "mon check distribution"
(which handling it properly wouldn't).

Signed-off-by: Andreas Ericsson <ae@op5.se>